refactor logging.go
